update utils to TS
