+++ src/utils/programUtils.ts
@@ -1,6 +1,8 @@
 import { TEST_USERS } from '../helpers';
 import { Program } from '../types';
-import { runGqlQuery } from './gatewayUtils';
+import { runGqlQuery, runGqlUpload, uploadFileFromString } from './gatewayUtils';
+import fs from 'fs';
+import urljoin from 'url-join';
 
 const generateProgram = (program?: Partial<Program>): Program => {
   const createTime = new Date();
@@ -29,32 +31,6 @@ const generateProgram = (program?: Partial<Program>): Program => {
       ],
     },
     ...program,
-  };
-};
-
-const generateProgram2 = () => {
-  const createTime = new Date();
-  const shortName = `Z${Math.floor(createTime.getTime() / 1000) % 10000000}-CA`;
-  return {
-    name: `Auto Generated Program ${shortName} - ${createTime.toISOString()}`,
-    shortName,
-    description: `This program was automatically generated for test purposes at ${createTime}`,
-    commitmentDonors: 4321,
-    website: 'https://exampletwo.com',
-    institutions: ['Toronto General Hospital'],
-    countries: ['Antarctica'],
-    regions: ['Africa', 'South America'],
-    cancerTypes: ['Multiple'],
-    primarySites: ['Stomach', 'Kidney'],
-    membershipType: 'ASSOCIATE',
-    admins: [
-      {
-        firstName: 'Oicr',
-        lastName: 'Testuser',
-        email: TEST_USERS.DCC_ADMIN.email,
-        role: 'ADMIN',
-      },
-    ],
   };
 };
 
@@ -87,4 +63,134 @@ const createProgram = ({ jwt, program }: { jwt: string; program: Program }) => {
   return runGqlQuery({ jwt, query, variables: { program } });
 };
 
-export { generateProgram, createProgram };
+const submitClinicalData = async ({
+  jwt,
+  shortName,
+  good,
+}: {
+  jwt: string;
+  shortName: string;
+  good: boolean;
+}) => {
+  const fileTypes = [
+    'donor',
+    'follow_up',
+    'hormone_therapy',
+    'primary_diagnosis-good',
+    'radiation',
+    'treatment',
+  ];
+
+  const query = `mutation ($shortName: String!, $files: [Upload!])
+  {uploadClinicalSubmissions (programShortName:$shortName, clinicalFiles:$files){
+    id
+    fileErrors {
+        message
+        fileNames
+        code
+      }
+  }}`;
+
+  const dataFiles = fileTypes.map(fileType => {
+    const folderPath = good ? './goodtestdata/' : './badtestdata/';
+    const filePath = urljoin(folderPath, `${fileType}.tsv`);
+    const file = fs
+      .readFileSync(filePath, 'utf8')
+      .split('DASH-CA')
+      .join(shortName);
+
+    return uploadFileFromString(file, fileType.concat('.tsv'));
+  });
+
+  return await runGqlUpload({
+    jwt,
+    query,
+    variables: { shortName },
+    files: dataFiles,
+  });
+};
+
+const registerSamples = async ({ jwt, shortName }: { jwt: string; shortName: string }) => {
+  const file = fs
+    .readFileSync('./goodtestdata/sample_registration.tsv', 'utf8')
+    .split('DASH-CA')
+    .join(shortName);
+
+  const query = `mutation ($files:Upload!, $shortName:String!) {
+    uploadClinicalRegistration(shortName:$shortName, registrationFile:$files) {
+      id
+      programShortName
+      creator
+      fileName
+      createdAt
+      records {
+        row
+        fields {
+          name
+          value
+        }
+      }
+      errors {
+        type
+        message
+        row
+        field
+        value
+        sampleId
+        donorId
+        specimenId
+      }
+      fileErrors {
+        message
+        fileNames
+        code
+      }
+      newDonors {
+        count
+        rows
+        names
+        values {
+          name
+          rows
+        }
+      }
+      newSpecimens {
+        count
+        rows
+        names
+        values {
+          name
+          rows
+        }
+      }
+      newSamples {
+        count
+        rows
+        names
+        values {
+          name
+          rows
+        }
+      }
+      alreadyRegistered {
+        count
+        rows
+        names
+        values {
+          name
+          rows
+        }
+      }
+    }
+  }`;
+
+  return await runGqlUpload({
+    jwt,
+    query,
+    variables: { shortName },
+    files: [uploadFileFromString(file, 'sample_registration.tsv')],
+    asArray: false,
+  });
+};
+
+export { generateProgram, createProgram, registerSamples, submitClinicalData };
